Allow castable values for convenience.

+++ src/test.cc
@@ -13,16 +13,16 @@ int main(void) {
 
   for (int i = 0; i < 10; i++) {
     // Immediate values.
-    cout << call(fd, inc, (int16_t)2) << ' ';
-    cout << call(fd, inc, (int16_t)4) << ' ';
-    cout << call(fd, inc, (int16_t)8) << ' ';
+    cout << call(fd, inc, 2) << ' ';
+    cout << call(fd, inc, 4) << ' ';
+    cout << call(fd, inc, 8) << ' ';
 
     // Variables.
     int16_t a = 1;
     cout << call(fd, inc, a) << '\n';
 
     // Void function.
-    call(fd, set_led, (uint8_t)63);
+    call(fd, set_led, 63);
   }
 
   close(fd);
